Add type inference to avoid warnings due to explicit casting (#1400)
* Add type inference to TestUtilities

* Add type inference to TransportConfig

* Add type inference to Deserializer

* Adjust generator classes for type inferring deserializer

* Add type inference to processNode to avoid casting when assigning results

+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/alarms/AlarmCommand.java
@@ -129,8 +129,8 @@ public class AlarmCommand extends ZclAlarmsCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        alarmCode = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
-        clusterIdentifier = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        alarmCode = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        clusterIdentifier = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
     }
 
     @Override
