+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/commissioning/ResetStartupParametersResponse.java
@@ -95,7 +95,7 @@ public class ResetStartupParametersResponse extends ZclCommissioningCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        status = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        status = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
     }
 
     @Override
