+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/metering/GetNotifiedMessage.java
@@ -182,9 +182,9 @@ public class GetNotifiedMessage extends ZclMeteringCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        notificationScheme = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
-        notificationFlagAttributeId = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        notificationFlagsN = (Integer) deserializer.deserialize(ZclDataType.BITMAP_32_BIT);
+        notificationScheme = deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        notificationFlagAttributeId = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        notificationFlagsN = deserializer.deserialize(ZclDataType.BITMAP_32_BIT);
     }
 
     @Override
